Fixed scheduler and executor
